==> _src/0.Data_import.ipynb <==
import pandas as pd
import psycopg2
from sqlalchemy import create_engine

## LOAD - Operations

## Load Operations 

operations_df = pd.read_csv('../_data/operation_pcd.csv')
operations_df.columns, operations_df.shape, operations_df.info()

test = operations_df[operations_df['subject_id']==121939743]
test

## Vitals
### LOAD - Vitals

## Load VITALS
vitals_df = pd.read_csv('../_data/vitals_in_hospital_filter.csv')
vitals_df.head()

test_vital = vitals_df[vitals_df['op_id']==494869962]
test_vital

### Pivot - Vitals

pivoted_vitals = pd.pivot_table(vitals_df, index=['op_id', 'subject_id', 'chart_time'], columns='item_name', values='value')
pivoted_vitals.head()
pivoted_vitals.info()


As a result of the pivot, these columns were converted into the index:  
- 0   #              5 non-null      int64  
- 1   **op_id**          5 non-null      int64  
- 2   **subject_id**     5 non-null      int64  
- 3   **chart_time**     5 non-null      int64  

Need to reset the index to the table can be used normallly again. 

# confirm that the index changed to the above: 
pivoted_vitals.index

#Reset the index! 
pivoted_vitals.reset_index(inplace=True)

# confim again. 
pivoted_vitals.index


## Test pivoted_vitals = ensure there remains only 1 record of an op_id
test_ppivot = pivoted_vitals[pivoted_vitals['op_id']==494869962]
test_ppivot


### Preliminary EDA - VITALS
1. Drop fields with high null count. 

vitals_todrop = ['bis','ci', 'ffp','ftn','pap_dbp','pap_mbp','pap_sbp','rbc','uo']
pivoted_vitals.drop(columns=vitals_todrop, inplace=True)
pivoted_vitals.info()

## LABS
### Load - Labs

## Load LABS
labs_df = pd.read_csv('../_data/labs_in_hospital_filter.csv')
labs_df.columns, labs_df.shape

### Pivot - Labs

pivoted_labs = pd.pivot_table(labs_df, index=['subject_id', 'chart_time'], columns='item_name', values='value')
pivoted_labs.head()

# confirm that the index changed to the above: 
pivoted_labs.index


#Reset the index! 
pivoted_labs.reset_index(inplace=True)

# confim again. 
pivoted_labs.index



# Testing: Because Labs does not have an OP_ID, we use Subject id instead. subject == 121939743

## Test pivoted_vitals = ensure there remains only 1 record of an op_id
testv_ppivot = pivoted_labs[pivoted_labs['subject_id']==121939743]
testv_ppivot

### Preliminary EDA - Labs
1. Drop fields with high NA

pivoted_labs.info()

labs_todrop = ['crp','hba1c']
pivoted_labs.drop(columns=labs_todrop, inplace=True)
pivoted_labs.info()

## MERGE 
### Operations_pcd with Vitals, then Labs.

operations_w_vitals_df = pd.merge(operations_df,pivoted_vitals, on='op_id', how='left')
operations_w_vitals_df.info()

operations_w_vitals_df.rename(columns={'subject_id_x': 'subject_id'}, inplace=True)
operations_w_vitals_df[operations_w_vitals_df['op_id']==494869962]

operations_vitals_labs_df = pd.merge(operations_w_vitals_df,pivoted_labs, on='subject_id', how='left')

operations_vitals_labs_df.shape, operations_vitals_labs_df.info()


## Export to CSV
### NOTES
* `operations_vitals_labs_df`contains duplicates due to joining LABS data. (no unique primary key between operations and labs)
* `operations_full_nodup` has had all duplicates (defined by: op_id / subject_id / hadm_id)

# EXPORT and Specify the file path 
output_csv_file = '../_data/operations_fulldata.csv'

# Save the DataFrame to a CSV file
operations_vitals_labs_df.to_csv(output_csv_file, index=False)

operations_full_nodup = operations_vitals_labs_df.drop_duplicates(subset=['op_id', 'subject_id', 'hadm_id'], keep='first')
operations_full_nodup

# EXPORT and Specify the file path 
output_csv_file = '../_data/operations_fulldata_nodup.csv'

# Save the DataFrame to a CSV file
operations_full_nodup.to_csv(output_csv_file, index=False)
