==> _src/02.EDA_2_LOS.ipynb <==
import pandas as pd
pd.options.display.float_format = '{:,.0f}'.format
# pd.options.display.int_format ='{:,}'.format

import os
from operator import itemgetter    
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings('ignore')
get_ipython().magic(u'matplotlib inline')
plt.style.use('ggplot')


import_csv_file = '../_data/operations_fulldata_nodup.csv'
df = pd.read_csv(import_csv_file)
df.info()

# calculate teh operation durations and add to DF
df['op_duration']=df['orout_time']-df['orin_time']
df['LOS']=(df['discharge_time']-df['orout_time'])/1440 # LOS is time following surgery to dishcarge converted to days. 
operations_info = df.loc[:,['op_id','category_id','category_desc','desc_short', 'icd10_pcs', 'orin_time','orout_time', 'op_duration', 'LOS']]

operations_info.info()


import pandas as pd

TOP_VAR_PROCEDURES = 100

# Group the DataFrame by 'icd10_pcs' and 'desc_schort' and calculate descriptive statistics for 'LOS'
descriptive_stats = operations_info.groupby(['category_id','category_desc'])['LOS'].describe()

# use the agg() method with a custom list of percentiles
percentiles = [0.1, 0.9]  # 
percentile_stats = operations_info.groupby(['category_id','category_desc'])['LOS'].quantile(percentiles).unstack()

# # Concatenate the percentile stats to the descriptive stats
descriptive_stats = pd.concat([descriptive_stats, percentile_stats], axis=1)

# # Rename the percentile columns for clarity
descriptive_stats = descriptive_stats.rename(columns={0.1: '10%', 0.9: '90%'})

# MODE or additional quantiles if needed
# descriptive_stats['mode'] = operations_info.groupby(['icd10_pcs','desc_short'])['LOS'].agg(pd.Series.mode)

operations_desc_stats_df = descriptive_stats.sort_values(by='std', ascending=False).head(TOP_VAR_PROCEDURES)
# descriptive_stats.to_csv('los.csv')

operations_desc_stats_df

## Clean up the index. 

operations_desc_stats_df.reset_index(inplace=True)
operations_desc_stats_df.index
operations_desc_stats_df


## Export List of PCD codes with greatest STD DEV to CSV

# EXPORT and Specify the file path 
output_csv_file = '../_data/category_STD.csv'

# Save the DataFrame to a CSV file
operations_desc_stats_df.to_csv(output_csv_file, index=False)

## Create a list of the surgeries with the greatest Variances in Length of stay
std_devs = operations_desc_stats_df['category_id']
print(std_devs)

# # Now, filter the DataFrame to these top 100 codes
top_LOSvariance_df = operations_info[operations_info['category_id'].isin(std_devs)]

top_LOSvariance_df



## Visualize LOS Distributions for surgeries

import plotly.graph_objects as go

import pandas as pd

df = top_LOSvariance_df

fig = go.Figure()

pcs_list = top_LOSvariance_df['category_id']

for pcs in pcs_list:
    fig.add_trace(go.Violin(x=df['category_id'][df['category_id'] == pcs],
                            y=df['LOS'][df['category_id'] == pcs],
                            name=pcs,
                            box_visible=True,
                            meanline_visible=True))

fig.show()
