==> _src/03.EDA_3_Scope.ipynb <==
import pandas as pd
pd.options.display.float_format = '{:,.0f}'.format
# pd.options.display.int_format ='{:,}'.format

import os
from operator import itemgetter    
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings('ignore')
get_ipython().magic(u'matplotlib inline')
plt.style.use('ggplot')


import_csv_file = '../_data/operations_fulldata_nodup.csv'
df = pd.read_csv(import_csv_file)
df.info()

import_csv_file = '../_data/category_STD.csv'
pcd_STD = pd.read_csv(import_csv_file)

import_csv_file = '../_data/category_VOL.csv'
pcd_VOL = pd.read_csv(import_csv_file)



pcd_STD_filter = pcd_STD['category_id']
pcd_VOL_filter = pcd_VOL['category_id']

pcd_inscope = pd.concat([pcd_STD_filter,pcd_VOL_filter],ignore_index=True)
pcd_inscope

## Create a filtered data set bason on the number of targeted pcs codes  

operations_inscope_df = df[df['category_id'].isin(pcd_inscope)]
operations_inscope_df.shape
# Next, group by the 'PCD_description' and count the occurrences
# description_counts = operartions_inscope_df.groupby('desc_short').size().sort_values(ascending=False)

operations_inscope_df

## Export IN_SCOPE to CSV for futher data preparation
operations_inscope_df.to_csv('../_data/operations_inscope_CATEGORIES.csv')


### Status Check:
* created a dynamic list of procedure codes based on surgery volume 
* created a second dynamic list of procedure coded based on LOS variance
* merged those lists to the 'in scope' category_id
* filtered the original dataset.
