==> _src/04.EDA_4_outliers_impute.ipynb <==
import pandas as pd


# Feature Engineering

df=pd.read_csv('../_data/operations_inscope_CATEGORIES.csv',index_col=0)
df.info()

## Identify LOS Outliers within each surgery type

We use `groupby`('icd10_pcs') groups the data by the unique values in the 'icd10_pcs'.  
 For each group, `transform` applies the calculate_percentile_90 function to values within that group.   

This group-specific 90th percentile is then used to determine if a value is an outlier within its category.
Lastly, the is_outlier column is created to indicate with 1 if the record is an outlier in its category or 0 otherwise. We also clean up by dropping the intermediate 'percentile_90' column.


import pandas as pd
import numpy as np

# Assuming df is your DataFrame, 'column_name' is the column you're examining,
# and 'category_column' is the column with the categorical data
df['LOS']=(df['discharge_time']-df['orout_time'])/1440 # LOS is time following surgery to dishcarge converted to days. 

# Define a function to calculate the 75th percentile for a series
def calculate_percentile_75(series):
    return series.quantile(0.90)

# Apply the function to each group's 'icd10_pcs' and create a new column for the 75th percentile
df['percentile_90'] = df.groupby('icd10_pcs')['LOS'].transform(calculate_percentile_75)

# Create a new column with 0/1 based on the condition, using the group-specific 75th percentile
df['is_outlier'] = np.where(df['LOS'] > df['percentile_90'], 1, 0)

# Drop the 'percentile_75' column if it's no longer needed
df.drop('percentile_90', axis=1, inplace=True)

df



### Classify if LOS was "Prolonged"
* set business rule as Prolonged is an LOS that is greater 1 std dev than the average. 

import pandas as pd
import numpy as np

# Define functions to calculate the mean and standard deviation for a series
def calculate_mean(series):
    return series.mean()

def calculate_std(series):
    return series.std()

# Apply the functions to each group's 'column_name' to create new columns for the mean and standard deviation
df['group_mean'] = df.groupby('icd10_pcs')['LOS'].transform(calculate_mean)
df['group_std'] = df.groupby('icd10_pcs')['LOS'].transform(calculate_std)

# Define one standard deviation away from the mean
df['mean_plus_std'] = df['group_mean'] + df['group_std']

# Create a new column with 0/1 based on the condition, 
# using the group-specific mean and standard deviation
df['prolonged_LOS'] = np.where(df['LOS'] > df['group_std'], 1, 0)

# Drop the 'group_mean' and 'group_std' columns if they're no longer needed
df.drop(['group_mean', 'group_std', 'mean_plus_std'], axis=1, inplace=True)
df
# Optionally, if you still need the filtered DataFrame without outliers, you can filter the DataFrame
# df_filtered = df[(df['prolonged_LOS'] == 1 )&(df['is_outlier'] == 0) ]
# df_filtered


df

## Identify Subject that had ICU visit (and duration)

## Create flag if subject went to ICU

df['icu_visit'] = (df['icuin_time'] > 0).astype(int)
df[df['icu_visit']== True]

# df['ICU_dur']= (df['icuout_time']-df['icuin_time'])
# Elected to omit ICR duration as most are Nan and woud increase complexity to drop later.

## 13771  subjects went to ICU post op. 



#########################
#
# Calculate OR Duration and Anesthetic Duration
#
#########################

#Operation Duration
df['or_duration']=(df['opend_time']-df['opstart_time'])

#Time under anasthetic
df['anesth_duration']=(df['anend_time']-df['anstart_time'])

df

## Drop extra column. Exclude Outliers from further analysis. 
cols_to_drop = ['is_outlier', 'art_dbp','chart_time_y','chart_time_x','inhosp_death_time','orin_time', 'orout_time', 'opstart_time', 'opend_time','admission_time', 'discharge_time','anstart_time' ,'anend_time', 'cpbon_time','cpboff_time','icuin_time','icuout_time', 'subject_id_y', 'chart_time_x','art_mbp','art_sbp','art_sbp','bt','cvp','alp','alt','ast', 'emop']
df = df[df['is_outlier'] == 0 ]
df = df.drop(columns=cols_to_drop)

df

df.info()

## Impute Medians for Missing Data
* cannot impute means across the whole data set (mix of male, female, age)
* means are imputed within each category (group be: sex, age, race)
* Note: this step is not performed in teh pipelines as important to find means for each 'group' of subject (ie: find mean platlet count for same gender, same age, same ASA)

## Confirm the column index that you want to include in the imput calcs.
df.info()

import pandas as pd
import numpy as np

# df_no_outlier = df[df['is_outlier'] == 0]
# Range to evaluate columns 32:58 (inclusive)

category_columns = ['age', 'sex', 'asa']  # asa is the id from Amercican Society of Anthropologists.
start_col = 17  # Start index 
end_col = 34    # End index (exclusive) 

# Iterate over each column in the specified range and apply groupby and transform
for col in df.iloc[:, start_col:end_col].columns:
    # Group by the categorical columns and fill NaN with the mean of the group for each specific column
    df[col] = df.groupby(category_columns)[col].transform(lambda x: x.fillna(x.mean()))

# Now df has the missing values imputed with the mean of their respective column within each group defined by 'category_columns'




df.info()

df = df.dropna()
df.shape


#########################
#
# Data EXport 
#
#########################
df.to_csv('../_data/operations_imputed_CLEAN_v2.csv')
